# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "This is the API documentation for Test Dalapa",
        "contact": {
            "email": "ulsyairil@outlook.co.id"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/ticket/verification": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Verifikasi status work order (disetujui/ditolak)",
                "operationId": "8ec9485e210071bd64748fb277833141",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "actor_id",
                                    "item_id",
                                    "status"
                                ],
                                "properties": {
                                    "actor_id": {
                                        "type": "string",
                                        "example": "user-123"
                                    },
                                    "item_id": {
                                        "type": "string",
                                        "example": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "DISETUJUI",
                                            "DITOLAK"
                                        ],
                                        "example": "DISETUJUI"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Verifikasi berhasil",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response sukses atau gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success or error message"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}